# api/openapi.json
{
  "components": {
    "schemas": {
...
      "httpx.ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/httpx.APIError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "media.DashboardItem": {
        "properties": {
          "channel_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "channel_name": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "mime_type": {
            "type": "string"
          },
          "original_filename": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "uploader_id": {
            "format": "uuid",
            "type": "string"
          },
          "uploader_name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "channel_id",
          "created_at",
          "id",
          "is_deleted",
          "mime_type",
          "original_filename",
          "size_bytes",
          "uploader_id",
          "uploader_name",
          "url"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "METRICS_TOKEN, only for /api/metrics.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "HttpOnly session cookie issued by POST /api/auth/login or /api/auth/register. The name is __Host-mnema_session when PUBLIC_URL is https.",
        "in": "cookie",
        "name": "mnema_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "REST API of Mnema Talk, a self-hosted, invite-only chat and voice app for a single community.\nAuthentication is an HttpOnly session cookie set by login/register. State-changing requests (anything but GET/HEAD/OPTIONS) must carry a same-origin Origin header, otherwise 403.\nJSON bodies are decoded strictly: unknown fields are rejected with 400 and bodies are limited to 1 MiB. Errors share the ErrorResponse shape.\nEverything except the SPA assets lives under /api.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit"
    },
    "title": "Mnema Talk API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/media": {
      "get": {
        "description": "Requires role admin (403 otherwise).",
        "operationId": "listMedia",
        "parameters": [
          {
            "description": "Page size. Values below 1 or non-numeric fall back to the default; larger values are clamped to 100.",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/media.DashboardItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Media page."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "FORBIDDEN: not allowed (admin required, not the author, wrong current password) or cross-origin request rejected by the CSRF check."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED: too many requests.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "List all media, newest first",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/media/prune": {
      "post": {
        "description": "Avatars are never pruned. Without days, uses MEDIA_RETENTION_DAYS; 400 when that is disabled. Requires role admin (403 otherwise).",
        "operationId": "pruneMedia",
        "parameters": [
          {
            "description": "Delete attachments older than this many days; defaults to MEDIA_RETENTION_DAYS.",
            "in": "query",
            "name": "days",
            "schema": {
              "maximum": 3650,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/media.PruneResult"
                }
              }
            },
            "description": "Prune result."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "Invalid input (INVALID_INPUT): malformed JSON, unknown JSON fields, bad IDs or failed validation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "FORBIDDEN: not allowed (admin required, not the author, wrong current password) or cross-origin request rejected by the CSRF check."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED: too many requests.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "UNAVAILABLE: a dependency (database, file storage) is not available."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete old chat attachments",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/media/stats": {
      "get": {
        "description": "Requires role admin (403 otherwise).",
        "operationId": "getMediaStats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/media.Stats"
                }
              }
            },
            "description": "Totals."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "FORBIDDEN: not allowed (admin required, not the author, wrong current password) or cross-origin request rejected by the CSRF check."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED: too many requests.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Storage totals",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/users/{userID}": {
      "get": {
        "description": "Includes voice_seconds and message_count. Other users are returned without the private presence choice.",
        "operationId": "getUser",
        "parameters": [
          {
            "description": "User ID.",
            "in": "path",
            "name": "userID",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.User"
                }
              }
            },
            "description": "The user."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "Invalid input (INVALID_INPUT): malformed JSON, unknown JSON fields, bad IDs or failed validation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "NOT_FOUND: the resource, or the route, does not exist."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED: too many requests.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get a user profile",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}